resturece to complete the splitting of parts
