+++ walter_modem/core.py
@@ -1,9 +1,6 @@
 import asyncio
 import time
 
-from machine import UART
-from .queue import Queue
-
 from .enums import (
     WalterModemOpState,
     WalterModemNetworkRegState,
@@ -17,8 +14,6 @@ from .enums import (
     WalterModemSimState,
     WalterModemHttpContextState,
     WalterModemSocketState,
-    WalterModemCMEErrorReportsType,
-    WalterModemCEREGReportsType,
     WalterModemMqttState,
     WalterModemMqttResultCode
 )
@@ -36,7 +31,6 @@ from .structs import (
     ModemGNSSAssistance,
     ModemCmd,
     ModemRsp,
-    ModemATParserData,
     ModemMqttMessage
 )
 
@@ -55,31 +49,31 @@ class ModemCore:
     SMALLER_THAN = ord('<')
     SPACE = ord(' ')
 
-    WALTER_MODEM_DEFAULT_CMD_ATTEMPTS = 3
+    DEFAULT_CMD_ATTEMPTS = 3
     """The default number of attempts to execute a command."""
 
-    WALTER_MODEM_PIN_RX = 14
+    PIN_RX = 14
     """The RX pin on which modem data is received."""
 
-    WALTER_MODEM_PIN_TX = 48
+    PIN_TX = 48
     """The TX to which modem data must be transmitted."""
 
-    WALTER_MODEM_PIN_RTS = 21
+    PIN_RTS = 21
     """The RTS pin on the ESP32 side."""
 
-    WALTER_MODEM_PIN_CTS = 47
+    PIN_CTS = 47
     """The CTS pin on the ESP32 size."""
 
-    WALTER_MODEM_PIN_RESET = 45
+    PIN_RESET = 45
     """The active low modem reset pin."""
 
-    WALTER_MODEM_BAUD = 115200
+    BAUD = 115200
     """The baud rate used to talk to the modem."""
 
-    WALTER_MODEM_CMD_TIMEOUT = 5
+    CMD_TIMEOUT = 5
     """The maximum number of seconds to wait."""
 
-    WALTER_MODEM_MIN_VALID_TIMESTAMP = 1672531200
+    MIN_VALID_TIMESTAMP = 1672531200
     """Any modem time below 1 Jan 2023 00:00:00 UTC is considered an invalid time."""
 
     MIN_PDP_CTX_ID = 1
@@ -88,35 +82,35 @@ class ModemCore:
     MAX_PDP_CTX_ID = 8
     """The highest possible PDP context ID"""
 
-    WALTER_MODEM_MAX_SOCKETS = 6
-    """The maximum number of sockets that the library can support."""
-
-    WALTER_MODEM_MAX_HTTP_PROFILES = 3
-    """The max nr of http profiles"""
-
-    WALTER_MODEM_MAX_TLS_PROFILES = 6
-    """The maximum number of TLS profiles that the library can support"""
-
-    WALTER_MODEM_OPERATOR_MAX_SIZE = 16
-    """The maximum number of characters of an operator name"""
-
-    WALTER_MODEM_MQTT_TOPIC_MAX_SIZE = 127
-    """The recommended mamximum number of characters in an MQTT topic"""
-
-    WALTER_MODEM_MQTT_MAX_PENDING_RINGS = 8
-    """The recommended maximum number of rings that can be pending for the MQTT protocol"""
-
-    WALTER_MODEM_MQTT_MAX_TOPICS = 4
-    """The recommended maximum allowed MQTT topics to subscribe to"""
-
-    WALTER_MODEM_MQTT_MIN_KEEP_ALIVE = 20
-    """The recommended minimum for the MQTT keep alive time"""
-
-    WALTER_MODEM_MQTT_MAX_MESSAGE_LEN = 4096
-    """The maximum MQTT payload length"""
-
     DEFAULT_PDP_CTX_ID = 1
     """The modem's default PDP CTX ID, if none is specified"""
+
+    MAX_SOCKETS = 6
+    """The maximum number of sockets that the library can support."""
+
+    MAX_HTTP_PROFILES = 3
+    """The max nr of http profiles"""
+
+    MAX_TLS_PROFILES = 6
+    """The maximum number of TLS profiles that the library can support"""
+
+    OPERATOR_MAX_SIZE = 16
+    """The maximum number of characters of an operator name"""
+
+    MQTT_TOPIC_MAX_SIZE = 127
+    """The recommended mamximum number of characters in an MQTT topic"""
+
+    MQTT_MAX_PENDING_RINGS = 8
+    """The recommended maximum number of rings that can be pending for the MQTT protocol"""
+
+    MQTT_MAX_TOPICS = 4
+    """The recommended maximum allowed MQTT topics to subscribe to"""
+
+    MQTT_MIN_KEEP_ALIVE = 20
+    """The recommended minimum for the MQTT keep alive time"""
+
+    MQTT_MAX_MESSAGE_LEN = 4096
+    """The maximum MQTT payload length"""
 
     def __init__(self):
         self._op_state = WalterModemOpState.MINIMUM
@@ -125,13 +119,13 @@ class ModemCore:
         self._reg_state = WalterModemNetworkRegState.NOT_SEARCHING
         """The current network registration state of the modem."""
 
-        self._socket_list = [ModemSocket(idx + 1) for idx in range(ModemCore.WALTER_MODEM_MAX_SOCKETS) ]
+        self._socket_list = [ModemSocket(idx + 1) for idx in range(ModemCore.MAX_SOCKETS) ]
         """The list of sockets"""
 
         self._socket = None
         """The socket which is currently in use by the library or None when no socket is in use."""
 
-        self._http_context_list = [ModemHttpContext() for _ in range(ModemCore.WALTER_MODEM_MAX_HTTP_PROFILES) ]
+        self._http_context_list = [ModemHttpContext() for _ in range(ModemCore.MAX_HTTP_PROFILES) ]
         """The list of http contexts in the modem"""
 
         self._http_current_profile = 0xff
@@ -292,7 +286,7 @@ class ModemCore:
                     self._add_at_byte_to_buffer(b, False)
 
                 elif self._parser_data.state == WalterModemRspParserState.DATA_HTTP_START2:
-                    if b == ModemCore.SMALLER_THAN and self._http_current_profile < ModemCore.WALTER_MODEM_MAX_HTTP_PROFILES:
+                    if b == ModemCore.SMALLER_THAN and self._http_current_profile < ModemCore.MAX_HTTP_PROFILES:
                         # FIXME: modem might block longer than cmd timeout,
                         # will lead to retry, error etc - fix properly
                         self._parser_data.raw_chunk_size = self._http_context_list[self._http_current_profile].content_length + len("\r\nOK\r\n")
@@ -370,7 +364,7 @@ class ModemCore:
 
             else:
                 tick_diff = time.time() - cmd.attempt_start
-                timed_out = tick_diff >= ModemCore.WALTER_MODEM_CMD_TIMEOUT
+                timed_out = tick_diff >= ModemCore.CMD_TIMEOUT
                 if timed_out or cmd.state == WalterModemCmdState.RETRY_AFTER_ERROR:
                     if cmd.attempt >= cmd.max_attempts:
                         if timed_out:
@@ -401,7 +395,7 @@ class ModemCore:
 
             else:
                 tick_diff = time.time() - cmd.attempt_start
-                if tick_diff >= ModemCore.WALTER_MODEM_CMD_TIMEOUT:
+                if tick_diff >= ModemCore.CMD_TIMEOUT:
                     await self._finish_queue_cmd(cmd, WalterModemState.TIMEOUT)
                 else:
                     return
@@ -447,7 +441,7 @@ class ModemCore:
         return WalterModemState.OK
 
     async def _handle_sqn_http_rcv_answer_start(self, tx_stream, cmd, at_rsp):
-        if self._http_current_profile >= ModemCore.WALTER_MODEM_MAX_HTTP_PROFILES or self._http_context_list[self._http_current_profile].state != WalterModemHttpContextState.GOT_RING:
+        if self._http_current_profile >= ModemCore.MAX_HTTP_PROFILES or self._http_context_list[self._http_current_profile].state != WalterModemHttpContextState.GOT_RING:
             return WalterModemState.ERROR
         else:
             if not cmd:
@@ -470,7 +464,7 @@ class ModemCore:
         http_status = int(http_status_str)
         content_length = int(content_length_str)
 
-        if profile_id >= ModemCore.WALTER_MODEM_MAX_HTTP_PROFILES:
+        if profile_id >= ModemCore.MAX_HTTP_PROFILES:
             # TODO: return error if modem returns invalid profile id.
             # problem: this message is an URC: the associated cmd
             # may be any random command currently executing */
@@ -497,7 +491,7 @@ class ModemCore:
         profile_id = int(profile_id_str)
         result_code = int(result_code_str)
 
-        if profile_id < ModemCore.WALTER_MODEM_MAX_HTTP_PROFILES:
+        if profile_id < ModemCore.MAX_HTTP_PROFILES:
             if result_code == 0:
                 self._http_context_list[profile_id].connected = True
             else:
@@ -508,7 +502,7 @@ class ModemCore:
     async def _handle_sqn_http_disconnect(self, tx_stream, cmd, at_rsp):
         profile_id = int(at_rsp[len("+SQNHTTPDISCONNECT: "):].decode())
 
-        if profile_id < ModemCore.WALTER_MODEM_MAX_HTTP_PROFILES:
+        if profile_id < ModemCore.MAX_HTTP_PROFILES:
             self._http_context_list[profile_id].connected = False
         
         return WalterModemState.OK
@@ -517,7 +511,7 @@ class ModemCore:
         profile_id_str, _ = at_rsp[len('+SQNHTTPSH: '):].decode().split(',')
         profile_id = int(profile_id_str)
 
-        if profile_id < ModemCore.WALTER_MODEM_MAX_HTTP_PROFILES:
+        if profile_id < ModemCore.MAX_HTTP_PROFILES:
             self._http_context_list[profile_id].connected = False
 
         return WalterModemState.OK
@@ -820,7 +814,7 @@ class ModemCore:
 
             if not first_key_parsed and len(pattern) > 2:
                 operator_name = pattern[:-2]
-                cmd.rsp.cell_information.net_name = operator_name[:ModemCore.WALTER_MODEM_OPERATOR_MAX_SIZE]
+                cmd.rsp.cell_information.net_name = operator_name[:ModemCore.OPERATOR_MAX_SIZE]
                 pattern = pattern[-2:]
                 first_key_parsed = True
 
@@ -1095,27 +1089,6 @@ class ModemCore:
                 if cur_cmd.state == WalterModemCmdState.COMPLETE:
                     cur_cmd = None
 
-    def _register_application_queue_rsp_handler(self, start_pattern: bytes, handler: callable):
-        if isinstance(start_pattern, bytes) and callable(handler):
-            if not self._application_queue_rsp_handlers_set:
-                self._application_queue_rsp_handlers_set = True
-                self._application_queue_rsp_handlers = [(start_pattern, handler)]
-            else:
-                self._application_queue_rsp_handlers.append((start_pattern, handler))
-        else:
-            log('WARNING', 'Invalid parameters, not registering application queue rsp handler')
-    
-    def _unregister_application_queue_rsp_handler(self, handler: callable):
-        if callable(handler):
-            if self._application_queue_rsp_handlers_set:
-                for i in range(len(self._application_queue_rsp_handlers) - 1, -1, -1):
-                    if self._application_queue_rsp_handlers[i][1] is handler:
-                        self._application_queue_rsp_handlers.pop(i)
-                if not self._application_queue_rsp_handlers:
-                    self._application_queue_rsp_handlers_set = False
-        else:
-            log('WARNING', f'Invalid paramater, cannot unregister: {type(handler)}, must be a callable')
-
     async def _run_cmd(self,
         at_cmd: str,
         at_rsp: str, 
@@ -1125,7 +1098,7 @@ class ModemCore:
         data = None,
         complete_handler = None,
         complete_handler_arg = None,
-        max_attempts = WALTER_MODEM_DEFAULT_CMD_ATTEMPTS
+        max_attempts = DEFAULT_CMD_ATTEMPTS
     ) -> bool:
         """
         Add a command to the command queue and await execution.
@@ -1174,38 +1147,3 @@ class ModemCore:
             cmd.rsp.result == WalterModemState.OK or
             (cmd.rsp.type == WalterModemRspType.HTTP and cmd.rsp.result == WalterModemState.NO_DATA)
         )
-
-    async def begin(self, debug_log: bool = False):
-        if not self._begun:
-            self.debug_log = debug_log
-            self._uart = UART(2,
-                baudrate=ModemCore.WALTER_MODEM_BAUD,
-                bits=8,
-                parity=None,
-                stop=1,
-                flow=UART.RTS|UART.CTS,
-                tx=ModemCore.WALTER_MODEM_PIN_TX,
-                rx=ModemCore.WALTER_MODEM_PIN_RX,
-                cts=ModemCore.WALTER_MODEM_PIN_CTS,
-                rts=ModemCore.WALTER_MODEM_PIN_RTS,
-                timeout=0,
-                timeout_char=0,
-                txbuf=2048,
-                rxbuf=2048
-            )
-
-            self._task_queue = Queue()
-            self._command_queue = Queue()
-            self._parser_data = ModemATParserData()
-
-            asyncio.create_task(self._uart_reader())
-            asyncio.create_task(self._queue_worker())
-
-            if not await self.reset():
-                raise RuntimeError('Failed to reset modem')
-            if not await self.config_cme_error_reports(WalterModemCMEErrorReportsType.NUMERIC):
-                raise RuntimeError('Failed to configure CME error reports')
-            if not await self.config_cereg_reports(WalterModemCEREGReportsType.ENABLED_UE_PSM_WITH_LOCATION_EMM_CAUSE):
-                raise RuntimeError('Failed to configure cereg reports')
-            
-            self._begun = True
